Refatorando lambda-function

+++ projeto/entities/product.py
@@ -1,7 +1,5 @@
-from decimal import Decimal
-
 class Product:
-    def __init__(self, id: str, name: str, author: str, rating: int, price: Decimal, file_name: str) -> None:
+    def __init__(self, id: str, name: str, author: str, rating: int, price: float, file_name: str) -> None:
         self.id = id
         self.name = name
         self.author = author
